refactoring check line friend bot api
