+++ src/api/Line/functions.ts
@@ -1,5 +1,8 @@
 import axios from "axios";
-import { GetLineNotificationStatusResponse } from "./types";
+import {
+  CheckLineBotFriendResponse,
+  GetLineNotificationStatusResponse,
+} from "./types";
 
 const ENDPOINT: string = import.meta.env.VITE_API_URL + "/api";
 
@@ -11,6 +14,15 @@ const getLineNotificationStatus =
     return res.data;
   };
 
+const checkLineBotFriend = async (): Promise<CheckLineBotFriendResponse> => {
+  const res = await axios.get(ENDPOINT + "/line/check-friend", {
+    withCredentials: true,
+  });
+
+  return res.data;
+};
+
 export const LineApi = {
   getLineNotificationStatus,
+  checkLineBotFriend,
 };
