+++ src/api/Line/types.ts
@@ -1,3 +1,7 @@
 export type GetLineNotificationStatusResponse = {
   is_notification: boolean;
 };
+
+export type CheckLineBotFriendResponse = {
+  friend_flag: boolean;
+};
