+++ src/api/User/hooks.ts
@@ -54,10 +54,3 @@ export const useCheckLogined = () => {
     queryFn: UserApi.checkLogined,
   });
 };
-
-export const useCheckLineBotFriend = () => {
-  return useQuery({
-    queryKey: ["users", "check-line-bot-friend"],
-    queryFn: UserApi.checkLineBotFriend,
-  });
-};
